Add WIDTHEXACT rule
There are many cases where users perceive WIDTH warnings too pedantic
and the sheer amount often leads to users using -Wno-WIDTH, thereby
hiding those that are more critical.

This adds a new rule WIDTHEXACT that is a style lint and thus not
enabled by default. It will contain WIDTH warnings that are strict in
expecting equal sizes, which is okay with the LRM, but can be a
mistake. WIDTH will probably contain mostly warnings that are even not
good with the LRM semantics (too large assignment, mismatch in
comparison, etc.)